Remove HashUsageChecker, use IScavengeMap instead.
HashInUseBefore used to just query a datastructure (intended: the index) to see how a hash is being used.
however we can't query the index efficiently to find out what we need

so we will have another datastructure to service this query instead.
however this datastructure needs to be constructed as well as queried.

turns out its just a map from hash to streamname, so we can use iscavengemap<ulong, string>

this also simplifies corner cases like
- reading from the index when some log entries have been scavenged,
- dealing with metadata for streams that dont exist
- removes worries about subtle ptable behaviours and old bugs like duplicate or out of order events